feat: Add Java 17 in most containers
All our Kotlin libraries are compiled to target JDK 17. Because they're not in the containers, they have to be downloaded each time from FooJay. However, FooJay is very unstable.

This will:
- Make images larger.
- Possibly increase download times when a job starts, due to increased image size.
- However, images should almost always already be downloaded on the runner. Downloading the JDK from FooJay isn't cached by the runner, so in the long run this may make jobs faster.

The notable exception to this is Prepared, which builds for JDK 11 and will still need to download a new JDK.